requirements.txt
flask_script
flask_migrate
nexmo